+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Section Order:
 - `tasks.send_fleet_invitation` refactored
 - `tasks.check_fleet_adverts` refactored
 - `tasks.get_fleet_composition` refactored
+- Fleet creation logic refactored and better error handling added
 
 ## [2.6.0] - 2025-08-14